+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-02-18
+
 ### Added
 
 - `#[format]` attribute for customizable field formatting
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for primitives, Option, HashMap, Vec, nested structs and `serde_json::Value`
 - Enable `preserve_order` feature for toml dependency to maintain field order
 
-[unreleased]: https://github.com/DiscreteTom/toml-scaffold/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/DiscreteTom/toml-scaffold/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/DiscreteTom/toml-scaffold/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/DiscreteTom/toml-scaffold/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/DiscreteTom/toml-scaffold/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/DiscreteTom/toml-scaffold/releases/tag/v0.1.0